Remove internal subroutine, add some print statements
I realized that some of the deaths were happening in the call to the
subroutine: gfortran dies with an error if you try to call a routine
with NaN, as makes sense - so, I'm getting rid of the subroutine call to
fix this. Also introducing some prints to better see what's going on.